+++ NVConso/PreferenceSection.cs
@@ -2,8 +2,6 @@ namespace NVConso
 {
     public enum PreferenceSection
     {
-        General,
-        Profiles,
         HeatMonitoring,
         History,
         Update,